# swagger/swagger-product.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "My User Project CRUD",
        "description": "My User Project Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": ["cnpmwarehouse.herokuapp.com"],
    "paths": {
        "/products": {
            "get": {
                "tags": ["Products"],
                "summary": "Get all product in system",
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer sYFuat5lz1y5v0LrCt7LfqJpo1AkdLgm7LbY6eRibN4NYw9Srf6aMIRJM8KbTwM6"
                        }
                    }, {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "string"

                        }

                    }, {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "string"

                        }

                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {

                            "$ref": "#/definitions/Products"
                        }
                    }

                }
            },
            "post": {
                "tags": ["Products"],
                "summary": "Import or Export product in system",
                "description": "Import or Export product in system",
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer sYFuat5lz1y5v0LrCt7LfqJpo1AkdLgm7LbY6eRibN4NYw9Srf6aMIRJM8KbTwM6"
                        }
                    }, {
                        "name": "product",
                        "in": "body",
                        "description": "product that we want to Import or Export",
                        "schema": {
                            "$ref": "#/definitions/AddProducts"
                        }
                    }, {
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer sYFuat5lz1y5v0LrCt7LfqJpo1AkdLgm7LbY6eRibN4NYw9Srf6aMIRJM8KbTwM6"
                        }
                    }
                ],
                "produces": ["application/json"],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/AddProducts"
                        }
                    },
                    "400": {
                        "description": "Cannot create product",
                        "schema": {
                            "$ref": "#/definitions/error400"
                        }
                    }
                }
            }

        },
        "/Products/{id}": {
            "parameters": [
                {
                    "in": "header",
                    "name": "Authorization",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "example": "Bearer sYFuat5lz1y5v0LrCt7LfqJpo1AkdLgm7LbY6eRibN4NYw9Srf6aMIRJM8KbTwM6"
                    }
                }, {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of user that we want to find",
                    "type": "integer"
                }
            ],
            "get": {
                "tags": ["Products"],
                "summary": "Get  product by ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Product"
                        }
                    }

                }
            }

        },
        "/products/categories": {
            "get": {
                "tags": ["Categories"],
                "summary": "Get all product by categories in system",
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer sYFuat5lz1y5v0LrCt7LfqJpo1AkdLgm7LbY6eRibN4NYw9Srf6aMIRJM8KbTwM6"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Categories"
                        }
                    }

                }
            },
            "post": {
                "tags": ["Categories"],
                "summary": "Create new categories in system",
                "description": "Create new categories in system",
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer sYFuat5lz1y5v0LrCt7LfqJpo1AkdLgm7LbY6eRibN4NYw9Srf6aMIRJM8KbTwM6"
                        }
                    }, {
                        "name": "category",
                        "in": "body",
                        "description": "Category that we want to create",
                        "schema": {
                            "$ref": "#/definitions/AddCategory"
                        }
                    }, {
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer sYFuat5lz1y5v0LrCt7LfqJpo1AkdLgm7LbY6eRibN4NYw9Srf6aMIRJM8KbTwM6"
                        }
                    }
                ],
                "produces": ["application/json"],
                "responses": {
                    "201": {
                        "description": "New category is created",
                        "schema": {
                            "$ref": "#/definitions/Category"
                        }
                    },
                    "400": {
                        "description": "Cannot create category",
                        "schema": {
                            "$ref": "#/definitions/error400"
                        }
                    }
                }
            }
        },
        "/Products/categories/{id}": {
            "parameters": [
                {
                    "in": "header",
                    "name": "Authorization",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "example": "Bearer sYFuat5lz1y5v0LrCt7LfqJpo1AkdLgm7LbY6eRibN4NYw9Srf6aMIRJM8KbTwM6"
                    }
                }, {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of category that we want to find product",
                    "type": "integer"
                }, {
                    "in": "query",
                    "name": "page",
                    "schema": {
                        "type": "string"

                    }

                }, {
                    "in": "query",
                    "name": "limit",
                    "schema": {
                        "type": "string"

                    }

                }
            ],
            "get": {
                "tags": ["Products"],
                "summary": "Get all product by cateroryID in system",
                "parameters": [
                    {
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Bearer sYFuat5lz1y5v0LrCt7LfqJpo1AkdLgm7LbY6eRibN4NYw9Srf6aMIRJM8KbTwM6"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {

                            "$ref": "#/definitions/Products"
                        }
                    }

                }
            }

        },
        "/Products/warehouse/{id}": {
            "parameters": [
                {
                    "in": "header",
                    "name": "Authorization",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "example": "Bearer sYFuat5lz1y5v0LrCt7LfqJpo1AkdLgm7LbY6eRibN4NYw9Srf6aMIRJM8KbTwM6"
                    }
                }, {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of warehouse that we want to find product",
                    "type": "integer"
                }, {
                    "in": "query",
                    "name": "page",
                    "schema": {
                        "type": "string"

                    }

                }, {
                    "in": "query",
                    "name": "limit",
                    "schema": {
                        "type": "string"

                    }

                }
            ],
            "get": {
                "tags": ["Products"],
                "summary": "Get all product by warehouseID in system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {

                            "$ref": "#/definitions/ProductByID"
                        }
                    }

                }
            }

        },
        "/Products/search/{productName}": {
            "parameters": [
                {
                    "in": "header",
                    "name": "Authorization",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "example": "Bearer sYFuat5lz1y5v0LrCt7LfqJpo1AkdLgm7LbY6eRibN4NYw9Srf6aMIRJM8KbTwM6"
                    }
                }, {
                    "name": "productName",
                    "in": "path",
                    "required": true,
                    "description": "Name of product to search",
                    "type": "string"
                }
            ],
            "get": {
                "tags": ["Products"],
                "summary": "Search products by name",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {

                            "$ref": "#/definitions/SearchProduct"
                        }
                    }

                }
            }

        }

    },
    "definitions": {
        "Product": {
            "required": [
                "name", "id", "categoryId"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "categoryId": {
                    "type": "integer"
                },

                "name": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "warehouse": {
                    "$ref": "#/definitions/Warehouse"
                }

            }
        },
        "Products": {
            "type": "array",
            "$ref": "#/definitions/Product"
        },
        "ProductByWarehouse": {
            "required": [
                "name", "id", "categoryId"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "categoryId": {
                    "type": "integer"
                },

                "name": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "WarehouseProduct": {
                    "$ref": "#/definitions/WarehouseProduct"
                }

            }
        },
        "WarehouseProduct": {
            "required": [
                "id", "warehouseId", "productId", "stock"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "warehouseId": {
                    "type": "integer"
                },
                "productId": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "Warehouse": {
            "required": [
                "name", "id", "cityId", "address"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "cityId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }

            }
        },
        "AddProduct": {
            "required": [
                "name", "categoryId"
            ],
            "properties": {

                "categoryId": {
                    "type": "integer"
                },

                "name": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "warehouseId": {
                    "type": "integer"
                },
                "stock": {
                    "type": "integer"
                },
                "actionType": {
                    "type": "string",
                    "example": "IMPORT||EXPORT"
                }

            }
        },
        "AddProducts": {
            "type": "object",
            "properties": {

                "products": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "Products1"
                            },
                            "categoryId": {
                                "type": "integer",
                                "example": 1
                            },
                            "stock": {
                                "type": "integer",
                                "example": 123
                            },
                            "warehouseId": {
                                "type": "integer",
                                "example": 141
                            },
                            "actionType": {
                                "type": "string",
                                "example": "IMPORT||EXPORT"
                            }
                        }
                    }

                }

            }
        },
        "Category": {
            "required": [
                "name", "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "products": {
                    "$ref": "#/definitions/Products"
                }

            }

        },
        "Categories": {
            "type": "array",
            "$ref": "#/definitions/Category"
        },
        "AddCategory": {
            "required": ["name"],
            "properties": {
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                }

            }

        },
        "ProductByCategory": {
            "required": [
                "name", "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "products": {
                    "$ref": "#/definitions/ProductWarehouse"
                }

            }

        },
        "ProductWarehouse": {
            "required": [
                "name", "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },

                "name": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                }

            }
        },
        "error400": {
            "properties": {
                "statusCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }

            }
        },
        "SearchProduct": {
            "required": ["name"],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }

            }
        }

    },
    "schemes": ["https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }

}